e2e tests
Reviewed-on: https://review.couchbase.org/c/perfrunner/+/191357
Tested-by: Build Bot <build@couchbase.com>

+++ perfrunner/remote/__init__.py
@@ -59,7 +59,10 @@ class Remote:
         with cd(worker_home):
             with cd('perfrunner'):
                 logger.info('move couchbase package to perfrunner')
-                run('mv /tmp/couchbase.rpm ./')
+                if os.path.isfile('/tmp/couchbase.rpm'):
+                    run('mv /tmp/couchbase.rpm ./')
+                else:
+                    run('mv /tmp/couchbase.deb ./')
 
     def start_celery_worker(self, worker, worker_home, broker_url):
         with settings(host_string=worker):
